first draft implementation of OpenMP - no gains in performance

+++ exercises/openmp/poisson/src/m_global.f90
@@ -16,9 +16,9 @@ MODULE m_global
    INTEGER              :: nx, ny, nz                ! cells per direction
    INTEGER              :: lx, ly, lz                ! box dimensions
    INTEGER              :: nsteps                    ! integration steps
-   REAL                 :: eps_min                   ! threshold on error value
-   !REAL                :: Tinit                     ! Tinit: initial field temperature
-   !REAL                :: Tboundary                 ! Tboundary: Dirilect-boundary field temperature
+   DOUBLE PRECISION                 :: eps_min                   ! threshold on error value
+   !DOUBLE PRECISION                :: Tinit                     ! Tinit: initial field temperature
+   !DOUBLE PRECISION                :: Tboundary                 ! Tboundary: Dirilect-boundary field temperature
    CHARACTER(LEN = 12)  :: source_type               ! source type, supports "radiator", "sinusoidal"
 
    ! ------------------------------------------------- !
@@ -32,7 +32,7 @@ MODULE m_global
    ! ------------------------------------------------- !
    ! DEPENDENT PARAMETERS                              !
    ! ------------------------------------------------- !
-   REAL      :: dx, dy, dz  ! dx: cell width
+   DOUBLE PRECISION      :: dx, dy, dz  ! dx: cell width
    
 END MODULE
